Implement prepend instrumentation for Curb (#577)
* set up prepend for curb

* updated config option

* update test for both chain and prepend

* update curb config

* updated curb intrumentation

* added missing file header

* updated chain instrumentation for curb

* added allowed_by_config? check for curb

* made instrumentation more DRY
